1. 리뷰 쓰기 추가 2. 404 Error page 추가 3. 카카오 결제완료 페이지 변경 4. 필요없는 데이터 삭제 및 자잘한 화면 변경

+++ user/models.py
@@ -42,7 +42,7 @@ class Bookmark(models.Model):
 
 class Reviews(models.Model):
     '''리뷰 테이블'''
-    star = models.FloatField(default=0.0)
+    star = models.IntegerField(default=0)
     review_content = models.TextField()
     review_data = models.DateField()
     # FK - 사용자
